Disables manufacturer field and adds year select
