+++ ppr-ui/src/components/mhrRegistration/YourHome/ManufacturerMakeModel.vue
@@ -18,6 +18,8 @@
               ref="nameRef"
               v-model="manufacturerName"
               filled
+              :disabled="isMhrManufacturerRegistration"
+              :class="{ 'unselectable' : isMhrManufacturerRegistration }"
               :rules="manufacturerNameRules"
               label="Business Legal Name"
               data-test-id="manufacturer-name"
@@ -25,58 +27,8 @@
           </v-col>
         </v-row>
 
-        <v-row id="mhr-home-manufacturer-year">
-          <v-col cols="2">
-            <label
-              class="generic-label"
-              for="manufacturer-year"
-              :class="{ 'error-text': validate && hasError(yearRef) }"
-            >
-              Year of Manufacture
-            </label>
-          </v-col>
-          <v-col cols="4">
-            <v-text-field
-              id="manufacturer-year"
-              ref="yearRef"
-              v-model="yearOfManufacture"
-              filled
-              :rules="manufactureYearRules"
-              label="Year of Manufacture"
-              persistent-hint
-              hint="YYYY"
-              data-test-id="manufacture-year"
-            />
-          </v-col>
-          <v-col cols="6">
-            <v-checkbox
-              id="circa-year"
-              label="This Year of Manufacture is approximate"
-              v-model="circa"
-              class="float-left"
-              data-test-id="circa-year-checkbox"
-            />
-            <v-tooltip
-              top
-              content-class="top-tooltip pa-5"
-              transition="fade-transition"
-              data-test-id="circa-year-tooltip"
-              nudge-right="3"
-              nudge-bottom="22"
-            >
-              <template v-slot:activator="{ on }">
-                <v-icon
-                  class="circa-tooltip-icon ml-2 mt-n1"
-                  color="primary"
-                  v-on="on"
-                  >mdi-information-outline</v-icon
-                >
-              </template>
-              When the exact year of manufacture is unknown, enter an estimated
-              year and indicate that the year is approximate.
-            </v-tooltip>
-          </v-col>
-        </v-row>
+        <ManufacturedYearInput v-if="!isMhrManufacturerRegistration" :validate="validate" />
+        <ManufacturedYearSelect v-else :validate="validate" />
 
         <v-divider class="mt-2 mb-5 mx-0 w-100" />
 
@@ -140,8 +92,11 @@ import { useStore } from '@/store/store'
 import { useInputRules, useMhrValidations } from '@/composables/'
 import { storeToRefs } from 'pinia'
 import { FormIF } from '@/interfaces'
+import ManufacturedYearInput from './ManufacturedYearInput.vue'
+import ManufacturedYearSelect from './ManufacturedYearSelect.vue'
 
 export default defineComponent({
+  components: { ManufacturedYearInput, ManufacturedYearSelect },
   props: {
     validate: {
       type: Boolean,
@@ -153,7 +108,6 @@ export default defineComponent({
     const makeModelComboForm = ref(null) as FormIF
     const nameRef = ref(null)
 
-    const yearRef = ref(null)
     const makeRef = ref(null)
     const modelRef = ref(null)
     const { setMhrHomeDescription, setMhrHomeBaseInformation } = useStore()
@@ -161,19 +115,15 @@ export default defineComponent({
       // Getters
       getMhrRegistrationValidationModel,
       getMhrRegistrationManufacturerName,
-      getMhrRegistrationYearOfManufacture,
-      getMhrRegistrationIsYearApproximate,
       getMhrRegistrationHomeMake,
+      isMhrManufacturerRegistration,
       getMhrRegistrationHomeModel
     } = storeToRefs(useStore())
     const {
       customRules,
       required,
-      minLength,
       maxLength,
-      startsWith,
-      greaterThan,
-      isNumber
+      greaterThan
     } = useInputRules()
 
     const {
@@ -182,17 +132,6 @@ export default defineComponent({
       hasError,
       setValidation
     } = useMhrValidations(toRefs(getMhrRegistrationValidationModel.value))
-
-    const manufactureYearRules = computed((): Array<Function> =>
-      customRules(
-        required('Enter a year of manufacture'),
-        isNumber(),
-        minLength(4, true),
-        maxLength(4, true),
-        startsWith(['19', '20'], 'Year must begin with 19 or 20'),
-        greaterThan(new Date().getFullYear() + 1, 'Year cannot be more than 1 year in the future')
-      )
-    )
 
     const combinedMakeModelLengthRule = (localState): Array<Function> => {
       return [
@@ -225,24 +164,12 @@ export default defineComponent({
     const localState = reactive({
       makeModelValid: false,
       manufacturerName: getMhrRegistrationManufacturerName.value,
-      yearOfManufacture: getMhrRegistrationYearOfManufacture.value?.toString(),
-      circa: getMhrRegistrationIsYearApproximate.value,
       make: getMhrRegistrationHomeMake.value,
       model: getMhrRegistrationHomeModel.value
     })
 
     watch(() => localState.manufacturerName, (val: string) => {
       setMhrHomeDescription({ key: 'manufacturer', value: val })
-    })
-
-    watch(() => localState.yearOfManufacture, (val: string) => {
-      if (parseInt(val)) {
-        setMhrHomeBaseInformation({ key: 'year', value: parseInt(val) })
-      }
-    })
-
-    watch(() => localState.circa, (val: boolean) => {
-      setMhrHomeBaseInformation({ key: 'circa', value: val })
     })
 
     watch(() => localState.make, (val: string) => {
@@ -264,16 +191,15 @@ export default defineComponent({
     return {
       hasError,
       nameRef,
-      yearRef,
       makeRef,
       modelRef,
       makeModelComboForm,
-      manufactureYearRules,
       makeRules,
       modelRules,
       manufacturerNameRules,
       maxLength,
       greaterThan,
+      isMhrManufacturerRegistration,
       ...toRefs(localState)
     }
   }
@@ -281,7 +207,7 @@ export default defineComponent({
 </script>
 
 <style lang="scss" scoped>
-::v-deep .circa-tooltip-icon {
-  line-height: 3em;
+::v-deep .unselectable {
+  user-select: none;
 }
 </style>
