+++ ppr-ui/src/components/mhrRegistration/YourHome/index.ts
@@ -4,3 +4,5 @@ export { default as HomeSections } from './HomeSections.vue'
 export { default as ManufacturerMakeModel } from './ManufacturerMakeModel.vue'
 export { default as RebuiltStatus } from './RebuiltStatus.vue'
 export { default as OtherInformation } from './OtherInformation.vue'
+export { default as ManufacturedYearInput } from './ManufacturedYearInput.vue'
+export { default as ManufacturedYearSelect } from './ManufacturedYearSelect.vue'
